Fix for FS#10881 - URL missing in some packages.
This patch adds URL to the fields to be updated on package update.

Path is a fix for the following scenario:
- a package is added at one point.
- later a url is changed or added.
- when the package is updated, the following fields are updated:
   pkgver, pkgrel, pkgdesc, needupdate, last_update

There was an original assumption, that some fields were very seldom changed,
and it would not have make sense from a performance standpoint to update the
fields that never changed, since the update mechanism was written using the
django orm..which is 'chatty'.

NOTE: It is rather pointless to do a compare to test if the field has changed.
It is more performant to simply update the field regardless.