Modify menu items order to be from top to bottom
The menu was rendered bottom to top in the html because the elements
were floated to the right. This caused text browsers visitors (no CSS
rendering) to see the menu backwards. I've modified the menu order and
made the items float to the left so the menu is correctly rendered in
text browsers.

Signed-off-by: Ismael Carnales <icarnales@gmail.com>
Signed-off-by: Dan McGee <dan@archlinux.org>