Touch up the depends and requiredby display in package details
This started out as a validation fix for the W3 validator: we had some <ul/>
elements that ended up on that page with no inner <li/> elements, so it was
invalid markup. I then realized we don't need to call the methods multiple
times so use the 'with' template tag.

Signed-off-by: Dan McGee <dan@archlinux.org>