Add OpenSearch support to the site
Implements FS#14185. It is a bit more complex than listed there as I wanted
to not hardcode the URLs in the descriptor file; to do this we need to make
it a template and fill some things in. We also need to serve the file using
the correct mime type.

Signed-off-by: Dan McGee <dan@archlinux.org>