Fix the get_user middleware stuff
First off, the copyright notice is from the other middleware previously
removed, so all of that can go, along with some stragglers in the import
list.

Next, as stated on
http://code.djangoproject.com/wiki/CookBookThreadlocalsAndUser, you need to
be sure to have the import in the model class be the same as the import used
in the middleware declaration, which was not true. Whoops.

Signed-off-by: Dan McGee <dan@archlinux.org>