reporead: allow a force update
This allows re-running repoadd on all packages in case of adding data or
fixing a bug without rendering the last_update values in the database
useless. For packages that aren't geting their version bumped, don't touch
last_update on a force import but do touch the rest of the fields.

Signed-off-by: Dan McGee <dan@archlinux.org>