Add CDN templatetag package
For now, this contains one new template tag- 'jquery'. This will allow us to
use the Google AJAX CDN in non-debug environments, since there is really no
need for us to be the source of this common file. In the future this package
may gain other static media resource tags as well.

Signed-off-by: Dan McGee <dan@archlinux.org>