Fix up HTML titles on a lot of pages
We didn't include them on many pages, or the ones we did weren't always
useful. Also try to keep the boilerplate to a minimum so you can see the
important bits in the title. 'Arch Linux - ' comes first in all titles, and
from there it can be filled in with something useful.

Signed-off-by: Dan McGee <dan@archlinux.org>