reporead: allow removal of last package in an arch
We had a situation where the last 'any' architecture package was present in
the [testing] repo and never got removed because we never did the
db_update() call on that architecture. Instead of looping all possible
architectures and only calling if len() > 0, always call db_update() for
both the primary architecture and the 'any' architecture.

Signed-off-by: Dan McGee <dan@archlinux.org>