Fix null field issues exposed by Django 1.1.2
Apparently Django 1.1.1 let null fields pass right through but this now
causes reporead to blow up in 1.1.2. Fix the issue and get things working
again by allowing nulls where it probably makes sense and including a
migration to fix the issue, which for the real database will be a no-op.

Signed-off-by: Dan McGee <dan@archlinux.org>