reporead: Add ability to generate traceback via signal
Every once in a while we see this command hanging on the main server but it
isn't making any system calls, so it is hard to tell where it is getting
stuck. Add a signal handler on SIGQUIT that will listen and print a
traceback when signaled.

This is the easiest thing to implement; future additions may need to be able
to hook up to a remote debugger (e.g. pdb) if this doesn't work.

Signed-off-by: Dan McGee <dan@archlinux.org>