Show maintainer full names in search dropdown
Instead of just showing usernames. It is especially helpful since we have
many similarly named developers at this point. In addition, only show active
developers/TUs and remove the rest from the list.

Also remove some cargo-cult code that set widget.choices; this is not
necesary as the widget automatically grabs the choices from the field.

Signed-off-by: Dan McGee <dan@archlinux.org>