Add links to 'Versions Elsewhere' on package details page
This will allow clicking through to a few more associated packages from a
given package, including those for the other architecture and potentially in
other repositories. The box floats to the right with the package actions,
which have also been restyled slightly so the whole area is a bit more
styled and set off from the rest of the page.

This should allow us to close the current oldest bug open in the Arch bug
tracker, FS#7787. I remember the days when every bug was a four-digit
number...

Signed-off-by: Dan McGee <dan@archlinux.org>