Split and update the pip requirements file
Add a second file for use in production environments where we know we
use memcached and MySQL, and for those that want to replicate the
production environment as close as possible. Slim down the primary file
to those requirements the app directly needs rather than how it is
installed. Also bump the version of South to the newer point release.

Signed-off-by: Dan McGee <dan@archlinux.org>