Developer dashboard fixes
As reported by Thomas, the dashboard JS dies when we try to apply table
sorting to a table with no entries (those in the top section). Use some
selector magic to only apply tablesorting if the table actually has rows
worth sorting.

Also move the package todo lists table down in the dashboard as the other
two tables are more relevant for the individual developer.

Signed-off-by: Dan McGee <dan@archlinux.org>