Adding changing of user profile details
The idea of this patch is allow to the developers who have an account, to
change their data without asking some admin to do it for them.

Dan: put private email address field back as it is used for a different
purpose; add some help text and field names as appropriate.

Signed-off-by: Angel Velasquez <angvp@archlinux.org>
Signed-off-by: Dan McGee <dan@archlinux.org>