Ensure mirrorlist generator works if no status available
Saw this error come through on the live site today, as well as being
reproducible when no mirror check runs have happened in the last 24 hours on
a development machine. Let mirrors that have no available checks show up on
this page, but be sorted last and show a score of unknown.

Signed-off-by: Dan McGee <dan@archlinux.org>