Slim down required by listings
If we have testing/non-testing packages in this list, and we are looking at
a package that is in both testing and non-testing, we can show only the
packages that correspond with the relevant repo.

I'm not sure any explanation will make this easier to understand, but the
end result is we don't show a bunch of duplicates where we used to.

Signed-off-by: Dan McGee <dan@archlinux.org>