Add developer world clocks page
This should make it easier for everyone to figure out whether someone is
awake, sleeping, and all that fun stuff. It does require everyone to update
their profile and fill in the field, but that shouldn't be too hard of a
task.

Don't suggest jokes to me Saturday morning on IRC unless you really want to
see them implemented. Thanks, Pierre!

Signed-off-by: Dan McGee <dan@archlinux.org>