Update tablesorter cache after clicking todo list status
We didn't update this before, so sorting the last after checking off a
few todos didn't work right, as it was using the old values. Fix it by
adding a call to update; this can eventually be changed to the
updateCell call commented out once I upgrade the minified tablesorter.js
file.

Signed-off-by: Dan McGee <dan@archlinux.org>