Slight tweaks to mirror commands
We know we are doing updates when setting IPv4/IPv6 information, so set
force_update to True to save the useless select query on each save(). For
mirror checks, use a less cumbersome deque for the results since it is also
thread-safe, and have all the log entries committed in one go.

Signed-off-by: Dan McGee <dan@archlinux.org>