Ensure package search form correctly handles errors
We were silently eating errors and just showing a normal package list if
the form didn't validate. Rather than do that, make sure we return no
packages at all and display the form errors back to the user in a sane
fashion. Adjust the validation methods on the 'limit' parameter so any
integer is acceptable.

Signed-off-by: Dan McGee <dan@archlinux.org>