Work out kinks in create new user view
We allowed repo selection before, but never actually called save_m2m()
so selections would have to be repeated on the next page. Add in group
selection as well to the form, and ensure we do all of our database
operations in one transaction so it is a lot easier to test and roll
back if things go wrong.

Signed-off-by: Dan McGee <dan@archlinux.org>