Utilize Django 1.3 'on_delete' feature on several foreign keys
The most important one here is PROTECT to keep people from making
bone-headed plays and deleting an Arch or Repo and every package along
with it. We can use this in a few other places, as well as some
carefully placed SET_NULL indicators.

Note that nothing here pushes deletion responsibilities down to the
database, although that will probably happen in a future commit.

Signed-off-by: Dan McGee <dan@archlinux.org>