Add new packages signoff model
This one is centered around pkgbase, much as our PackageRelation object
is. However, it also tracks all of the versioning fields we have in
order to making joining against the current package testing list
possible. Finally, additional metadata including a created date, an
(optional) revoke date, and a comments field are added.

Signed-off-by: Dan McGee <dan@archlinux.org>