templates/flag.html: Improve the note about bug reports.
There are tons of morons out there who are not able to read. Thus, I
keep getting bug reports via the "flag out of date" feature.
This patch removes all occurrences of "please" from the note and
adds a new sentence that explains our ignorance properly. The hint
to file a bug report is now longer, so idiots might see it.

Dan: add one more line in the early text.

Signed-off-by: Dan McGee <dan@archlinux.org>