Add some dev dashboard info regarding signed package count
This adds a column similar to the flagged package count for the number
of signed packages in a given architecture or repository. It is up to
the user to do some simple math to figure out the number of unsigned
packages.

Also, add 'signed' as a hidden search field option similar to what we
did for packager.

Signed-off-by: Dan McGee <dan@archlinux.org>