Store flag requests in the database
This makes them persistent rather than the transient beings they
currently are. We attempt to capture all the metadata we need to be able
to do things with this later- aka IP address (for spam checking later),
fields that allow us to mark the request as spam or not an actual
out-of-date report, etc.

As a bonus, logged-in developers now get the email address field filled
in for free. Yay.

Signed-off-by: Dan McGee <dan@archlinux.org>