Add old version string to saved flag requests
This makes it easier to match up a flag request with the package state
at the time of flagging, and might also help to determine if flagging
actions were legit. We only store it if it is the same across all
packages to be marked.

Also, move the various database write activities when flagging packages
into a single transaction.

Signed-off-by: Dan McGee <dan@archlinux.org>