Extract split_package_details() method
This is never currently called directly as a view method, but is used by
the normal package details view as a fallback if a package cannot be
located.

This also fixes an issue where looking up a package in a repo it is not
in returns the split details page for one package, which is incorrect
behavior.

Signed-off-by: Dan McGee <dan@archlinux.org>