Make sorting by time on clocks page more predictable
This is a little bit of a hack but works well. Add the timezone name to
the end of the sort so those in the same time zone end up next to each
other. For timezones like CEST that have many different specifiers, it
makes more sense to group them by 'Europe/Berlin', 'Europe/Paris', etc.

Signed-off-by: Dan McGee <dan@archlinux.org>