Flagging related cleanups and improvements
Touch up the style slightly on the flag help popup to match the main
site style more closely.

When a logged-in user is flagging a package out of date, we have no need
for them to fill in the email field since we already have an email
address on file.

Signed-off-by: Dan McGee <dan@archlinux.org>