Collapse the dependencies and required by lists when they are long
For now, this happens when the lists are over 20 items. Using JS, hide
the 21st and following packages listed in the list and replace them with
a 'Show More...' link that users can click to get the full list.

For a package such as glibc with 444 'Required By' entries, this can
make quite a visual difference.

Signed-off-by: Dan McGee <dan@archlinux.org>