Revert "Fall back to 410 Gone for package files view as well"
This reverts commit 9ab460c53a1ac4c79da6f05f2850ee21beedbab2.

This seemed like the right thing to do, but it doesn't really play well
with our more general dispatch framework we now do on the package
details pages. Just let it 404 like it always did, as these pages are
less essential. We can perhaps add a full dispatcher later if we really
feel the need.