Add some HTML5-ization in JS of various input attributes
On the login page, give focus to the username box when the page loads as
well as turning autocorrection and auto-capitalization off on the
username box.

For the developer profile page, we can add some minor validation and
typing of certain form fields that allow things like iPhone and Android
to customize the presented keyboard to the user, as well as allowing
browsers to do some client-side validation.

Signed-off-by: Dan McGee <dan@archlinux.org>