Explicitly close the database connection in reporead
This is the cause of these warnings showing up in the PostgreSQL log:
    LOG:  unexpected EOF on client connection with an open transaction

All management commands are guilty of this as they do not clean up and
close the connection when they exit, unlike the standard web request
cycle. Other commands should probably be updated as well, but for now,
this is the biggest culprit.

Signed-off-by: Dan McGee <dan@archlinux.org>