Don't cache package properties as aggressively
For package signatures, it turns out it is way cheaper to just parse the
signature again rather than going though all the decorator and
cache_function_key business. This speeds up the mismatched signatures
report significantly once this is removed.

For base_package, given that we only call it once from our package
details template, it makes little sense to cache the result.

Signed-off-by: Dan McGee <dan@archlinux.org>