Add new todolists and todolist package model
Move the todolist model from main to the todolists application, and make
a few minor tweaks to field names along the way. Also add a 'raw' field
that will hold the originally input text data from the creator or last
modifier of the todolist.

Add pkgname, pkgbase, arch, and repo fields to a new todolist package
model, which will supplement the former foreign key to an actual package
object. This will prevent todolist package objects from ever being
deleted as they can be now, which is not intuitive.

Also change the current boolean 'complete' flag to a 'status' enum that
can hold other values. For now, we add 'In-progress' to the mix.

Finally, add a 'user' field, and a 'comments' field that will be
utilized later by the UI.

Signed-off-by: Dan McGee <dan@archlinux.org>