Add data migration for todolists
This moves the data from the old models into the new ones. Note that IDs
are not preserved across the move, but we do store the old ID in the
old_id column so we don't break every link out there.

Links will become slugs in a future commit, so there should be no
ambiguity when linking via number vs string.

Signed-off-by: Dan McGee <dan@archlinux.org>