Clean up and make several migrations modern
This moves most migrations to the v2 format that have been presenting
some issues. One missing depends_on relationship has been added, and we
allow an index to not be dropped if it does not exist due to the
shittyness in sqlite3 actually keeping indexes across DDL on that table.

Signed-off-by: Dan McGee <dan@archlinux.org>