Don't pull and sort mirror URLs unless we have to
On the download page, the explicit sorted() call was forcing evaluation
of the Django queryset, even if we never actually needed the results
because the template fragment was cached. Wrap it all in a callable
function which looks the same to the template, but saves us the cost of
evaluation every single page view.

Signed-off-by: Dan McGee <dan@archlinux.org>