Use a subquery rather than two queries in attach_maintainers
Now that we are using a database that doesn't stink, it makes more sense
to do all of the stuff we need to do down at the database level. This
helps a lot when 500+ packages are in play at a given time, such as
some of our larger rebuild todo lists.

Signed-off-by: Dan McGee <dan@archlinux.org>